perf tools: Add new Intel PT packet definitions
New features have been added to Intel PT which include a number
of new packet definitions.

This patch adds packet definitions for new packets: TMA, MTC, CYC,
VMCS, TRACESTOP and MNT.  Also another bit in PIP is defined.

This patch only adds support for the definitions. Later patches
add support for decoding TMA, MTC, CYC and TRACESTOP which is
where those packets are explained.

VMCS and the newly defined bit in PIP are used with virtualization
which is not supported yet.  MNT is a maintenance packet which the
decoder should ignore.

For details, refer to the June 2015 or later Intel 64 and
IA-32 Architectures SDM Chapter 36 Intel Processor Trace.

Signed-off-by: Adrian Hunter <adrian.hunter@intel.com>