perf tools: Pass Intel PT information for decoding MTC and CYC
Record additional information in the AUXTRACE_INFO
event in preparation for decoding MTC and CYC packets.
Pass the information to the decoder.

The AUXTRACE_INFO record can be extended by using the
size to indicate the presence of new members.

The additional information includes PMU config bit
positions and the TSC to CTC (hardware crystal clock)
ratio needed to decode MTC packets.

Signed-off-by: Adrian Hunter <adrian.hunter@intel.com>