perf tools: Validate config term maximum value
Currently the value of a PMU config term is silently
truncated if it is too big. This is an impediment to
validating the value for other criteria later on i.e.
the user provides an invalid value that gets truncated
to a valid one.

The maximum value validation is only done for the
parser where the error is passed back to the user. In
other cases the silent truncation continues so as not
to affect tools that perhaps rely on it.

Signed-off-by: Adrian Hunter <adrian.hunter@intel.com>