perf tools: Add Intel PT support for decoding TRACESTOP packets
A TRACESTOP packet is produced when an Intel PT trace
enters a defined region of the address space at which
point the tracing stops.

This patch just adds decoder support.

Support for specifying TRACESTOP regions is left until
later.

For details refer to the June 2015 or later Intel 64 and
IA-32 Architectures SDM Chapter 36 Intel Processor Trace.

Signed-off-by: Adrian Hunter <adrian.hunter@intel.com>